fix(stackmap): 保真编码第五类寄存器根掩码
- runtime/stackmap_codec.rs 的 map record 头原本只写四类 u16 掩码并补 4 字节
  保留位，第五类 StackInterior 掩码被静默丢弃；编码器现按规范写入五个掩码，
  仅保留最后 2 字节 reserved
- verify_map_record 同步读取五个掩码，校验 bit15 为零、reserved 为零、五类
  互斥，并新增 lane 上界拒绝
- 新增 codec_preserves_fifth_register_root 回归：直接断言 map data 偏移处的
  第五类掩码字节，并验证与 direct 掩码重叠时必须拒绝
- 清理栈图子系统既有 clippy 告警（identity_op、map_identity、needless_range_loop、
  manual_is_multiple_of、useless_conversion、collapsible_if）

+++ crates/gugu-compiler/src/runtime/stackmap.rs
@@ -214,7 +214,7 @@ pub(crate) fn scan_roots(
             continue;
         }
         let cage = (word >> 56) as u8;
-        let generation = ((word >> 32) & 0xff_ffff) as u64;
+        let generation = (word >> 32) & 0xff_ffff;
         let offset_value = word & 0xffff_ffff;
         let descriptor = world
             .cages
